[WIP] got checkboxes partially styled, not completely
