+++ components/Checkbox/Checkbox.scss
@@ -1,5 +1,67 @@
 @import '../util/include-responsive.scss';
 
+.standardCheckboxWrapper {
+  $paddingProps: (
+    orientation: 'bottom-only',
+    bottom-side: 15px,
+  );
+
+  @include Padding($paddingProps);
+  height: 20px;
+  width: 20px;
+}
+
+.standardCheckbox {
+  width: 20px;
+  position: relative;
+
+  input[type=checkbox]{
+    opacity: 0;
+    position: absolute;
+     &:checked{
+      + label{
+        background-color: $VimeoBlue;
+        & :after {
+         opacity: 1;
+        }
+      }
+    }
+  }
+}
+
+.standardCheckboxLabel {
+  cursor: pointer;
+  position: absolute;
+  width: 20px;
+  height: 20px;
+  background: $White;
+  border:1px solid $Porcelain;
+  border-radius: 2px;
+
+  &:hover {
+    background-color: $VimeoBlue;
+    &::after {
+      opacity: 1;
+    }
+  }
+
+  &:after {
+    opacity: 1;
+    content: '';
+    position: absolute;
+    width: 13px;
+    height: 8px;
+    background: transparent;
+    top: 3px;
+    left: 3px;
+    border: 3px solid $White;
+    border-top: none;
+    border-right: none;
+
+    transform: rotate(-45deg);
+  }
+}
+
 .checkbox {
   transition: border 400ms ease-out;
   border: 1px solid $Porcelain;
